+++ src/code_context_analyzer/analyzer/parsers/__init__.py
@@ -3,6 +3,6 @@ from .js_parser import JSParser
 from .python_parser import PythonParser
 
 registry = {
-    'py': PythonParser(),
-    'js': JSParser(),
+    'python': PythonParser(),
+    'javascript': JSParser(),
 }